Use BIGINT in MySQL which fits unix timestamp for Bug#966562

+++ docs/db/schema_up_000.sql
@@ -2,7 +2,7 @@ CREATE DATABASE apk_factory CHARACTER SET utf8 COLLATE utf8_general_ci;
 USE apk_factory;
 CREATE TABLE apk_metadata (
   id CHAR(40) NOT NULL,
-  version INT,
+  version BIGINT,
   manifest_url VARCHAR(5000),
   manifest_hash CHAR(40),
   package_hash CHAR(40),
